refactor: 도피와 하디 등록뷰 작업 합침
