+++ GilCat/GilCat/Sources/Global/Extensions/ColorExtension.swift
@@ -4,7 +4,5 @@ extension Color {
     static let backgroundColor = Color("BackGroundColor")
     static let buttonColor = Color("ButtonColor")
     static let pickerColor = Color("PickerColor")
-    static let lightOrange = Color("ButtonColor")
-    static let darkBlue = Color("DarkBlueColor")
-    static let darkGray = Color("GrayColor")
+    static let profileBackgroundColor = Color("GrayColor")
 }
