+++ GilCat/GilCat/Sources/Screen/Registers/registerViews/register2.swift
@@ -37,12 +37,12 @@ struct Register2: View {
                 Button {
                 
                 } label: {
-                    CustomMainButton(text: "건너뛰기", foreground: Color.white, background: Color.darkBlue)
+                    CustomMainButton(text: "건너뛰기", foreground: .white, background: .pickerColor)
                 }
                 Button {
                 
                 } label: {
-                    CustomMainButton(text: "시작하기", foreground: Color.white, background: Color.lightOrange)
+                    CustomMainButton(text: "시작하기", foreground: .white, background: .buttonColor)
                 }
             }
             .padding()
@@ -58,7 +58,7 @@ struct Register2: View {
     func getCodeInputView(index: Int) -> some View {
         return TextEditor(text: $codeInput[index])
             .frame(width: 50, height: 65)
-            .background(Color.darkBlue)
+            .background(Color.pickerColor)
             .cornerRadius(20)
             .multilineTextAlignment(.center)
             .foregroundColor(Color.white)
